content_aggregator/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-crispy-forms==1.10.0
django-filter==2.4.0
djangorestframework==3.12.2
html5lib==1.1
idna==2.10
importlib-metadata==3.4.0
Markdown==3.3.3
Pillow==8.1.0
pytz==2020.5
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
webencodings==0.5.1
zipp==3.4.0
psycopg2-binary==2.8
